requirements.txt
streamlit
google-generativeai
python-dotenv
streamlit-js-eval  # <-- ADD THIS LINE
